patch: Fixing circular dependency in integration tests
