+++ actions.yaml
@@ -97,3 +97,12 @@ force-refresh-start:
 
         If `false`, force refresh if new version of OpenSearch and/or charm is not compatible with previous version
   additionalProperties: false
+
+status-detail:
+  description: Gets statuses of the charm
+  params:
+    recompute:
+      type: boolean
+      default: false
+      description: a boolean indicating whether a unit should recompute all statuses.   - continue upgrade if 1+ upgraded units have non-active status
+
