+++ metadata.yaml
@@ -20,6 +20,8 @@ peers:
     # DA056 - Upgrading in-place upgrade protocol
     # https://docs.google.com/document/d/1H7qy5SAwLiCOKO9xMQJbbQP5_-jGV6Lhi-mJOk4gZ08/edit
     interface: upgrade
+  status-peers:
+    interface: status-peers
 
 provides:
   peer-cluster-orchestrator:
