+++ tests/integration/helpers.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
-import asyncio
 import json
 import logging
 import random
@@ -23,7 +22,6 @@ from tenacity import (
     wait_random,
 )
 
-from .continuous_writes import ContinuousWrites
 from .helpers_deployments import get_application_units
 
 METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
@@ -118,54 +116,6 @@ async def get_shards_by_index(ops_test: OpsTest, unit_ip: str, index_name: str) 
 
     return result
 
-
-async def assert_continuous_writes_increasing(
-    c_writes: ContinuousWrites,
-) -> None:
-    """Asserts that the continuous writes are increasing."""
-    writes_count = await c_writes.count()
-    await asyncio.sleep(20)
-    more_writes = await c_writes.count()
-    assert more_writes > writes_count, "Writes not continuing to DB"
-
-
-async def assert_continuous_writes_consistency(
-    ops_test: OpsTest, c_writes: ContinuousWrites, apps: List[str]
-) -> None:
-    """Continuous writes checks."""
-    result = await c_writes.stop()
-    logger.info(f"Continuous writes result: {result}")
-    assert result.max_stored_id == result.count - 1
-    assert result.max_stored_id == result.last_expected_id
-
-    unit_ip = await get_leader_unit_ip(ops_test, apps[0])
-
-    # fetch unit ips by unit id by application
-    apps_units_ips = {app: await get_application_unit_ids_ips(ops_test, app) for app in apps}
-
-    # investigate the data in each shard, primaries and their respective replicas
-    shards = await get_shards_by_index(ops_test, unit_ip, ContinuousWrites.INDEX_NAME)
-    shards_by_id = {}
-    for shard in shards:
-        shards_by_id.setdefault(shard.num, []).append(shard)
-
-    # count data on each shard. For the **balanced** continuous writes index, we have 2
-    # primary shards and replica shards of each on all the nodes. In other words: prim1 and
-    # its replicas will have a different "num" than prim2 and its replicas.
-    count_from_shards = 0
-    for shard_num, shards_list in shards_by_id.items():
-        count_by_shard = [
-            await c_writes.count(
-                unit_ip=apps_units_ips[shard.app][shard.unit_id],
-                preference=f"_shards:{shard_num}|_only_local",
-            )
-            for shard in shards_list
-        ]
-        # all shards with the same id must have the same count
-        assert len(set(count_by_shard)) == 1
-        count_from_shards += count_by_shard[0]
-
-    assert result.count == count_from_shards
 
 
 @retry(wait=wait_fixed(wait=15), stop=stop_after_attempt(15))
